fixes check if identity is empty
